fix: 팝업 로그인 관련 에러사항 수정
