--- apps/extension/hooks/useSearch/index.ts
@@ -1,1 +1,0 @@
-export { useSearch } from './useSearch'
