+++ apps/web/src/app/layout.tsx
@@ -8,7 +8,6 @@ import { pretendard } from '@/styles/fonts'
 import { MSWComponent } from '@/mock/MSWComponent'
 import { ToastContextProvider } from '@/hooks/useToast'
 import { InitialSetting } from '@/components/InitialSetting'
-import TokenSender from '@/components/TokenSender/TokenSender'
 
 import { ModalContextProvider } from 'src/hooks/useModal/useModal'
 
@@ -27,7 +26,6 @@ const RootLayout = ({
       <ModalContextProvider>
         <ToastContextProvider>
           <InitialSetting />
-          <TokenSender />
           <div>{children}</div>
           <div id="modal" />
         </ToastContextProvider>
