+++ apps/web/src/components/InitialSetting/InitialSetting.tsx
@@ -14,10 +14,6 @@ export const InitialSetting = () => {
       location.href = '/login'
     })
     baseFetcher.setErrorHandler(() => {
-      // addToast({
-      //   message: error.message,
-      //   type: 'error',
-      // })
       addToast({
         message: '죄송합니다. 오류가 발생했습니다.\n잠시 후 다시 시도해주세요.',
         type: 'error',
